Add review scraping possibility for later use
NOTE: Currently only supports Kaaoszine.fi urls for scraping. Plan is to
add support for more sites later, that's part of why the domain extraction
part was added in the first place. Later it will be leveraged so I know
which functions to use to extract the data I need.

It will create the complete row and also the reference part, defaulting to
the current day for when the page was visited.