Fix bugged Kaaoszine reviews. Add support for Soundi.
Since Soundi release dates are formatted in only month/year
format, it was necessary to add another constant variable for
the month part so the reference looks good. Not thrilled with
the solution but it'll work for now, maybe in the future one
could use a single list and use title on it and add 'ta' to the
end of the month if needed instead of having two separate lists.